Fixed some issues with database and pages. And fixed analysis static url

+++ web/src/classes/Database/DatabaseSession.php
@@ -60,8 +60,7 @@ class DatabaseSession
                     ".time().")";
         $result = $mysqli->query($sql);
 
-        if(!$result)
-            return null;
+        if(!$result) return null;
 
         // Get the session ID
         $sessionID = Database::safe($mysqli->insert_id, $mysqli);
@@ -79,6 +78,8 @@ class DatabaseSession
                 $sql = "INSERT INTO `yacrs_sessionsAdditionalUsers` (`sessionID`, `userID`)
                         VALUES ('$sessionID', '$additionalUserId'); ";
                 $result = $mysqli->query($sql);
+
+                if(!$result) return null;
             }
         }
 
@@ -87,8 +88,7 @@ class DatabaseSession
                 VALUES ('$sessionID') ";
         $result = $mysqli->query($sql);
 
-        if(!$result)
-            return null;
+        if(!$result) return null;
 
         // Get the session identifier
         $sessionIdentifier = Database::safe($mysqli->insert_id, $mysqli);
@@ -165,6 +165,8 @@ class DatabaseSession
                 $sql = "INSERT INTO `yacrs_sessionsAdditionalUsers` (`sessionID`, `userID`)
                         VALUES ('$sessionID', '$additionalUserId'); ";
                 $result = $mysqli->query($sql);
+
+                if(!$result) return null;
             }
         }
         return true;
@@ -303,13 +305,11 @@ class DatabaseSession
                       AND sau.`userID` = u.`userID`";
             $result2 = $mysqli->query($sql);
 
-            // If query was successful
-            if($result2) {
+            if(!$result2) return null;
 
-                // Loop for every row and add additional user
-                while($row2 = $result2->fetch_assoc()) {
-                    $session->addAdditionalUser($row2["username"]);
-                }
+            // Loop for every row and add additional user
+            while($row2 = $result2->fetch_assoc()) {
+                $session->addAdditionalUser($row2["username"]);
             }
 
             // Add session to array of sessions
@@ -342,6 +342,9 @@ class DatabaseSession
                 FROM `yacrs_sessionIdentifier`
                 WHERE `sessionIdentifier` = $sessionIdentifier";
         $result = $mysqli->query($sql);
+
+        if(!$result) return null;
+
         $row = $result->fetch_assoc();
         $sessionID = $row["sessionID"];
 
@@ -351,32 +354,33 @@ class DatabaseSession
                 WHERE `sessionID` = $sessionID";
         $result2 = $mysqli->query($sql);
 
-        // If query was successful
-        if($result2) {
-            // Loop for every row
-            while($row2 = $result2->fetch_assoc()) {
-                $questionID = $row2["questionID"];
-                DatabaseSessionQuestion::delete($questionID, $mysqli);
-                //Delete from questions
-                $sql = "DELETE FROM `yacrs_questions`
-                        WHERE `questionID` = $questionID";
-                $result = $mysqli->query($sql);
-            }
-        }
+        if(!$result2) return null;
 
+        // Loop for every row
+        while($row2 = $result2->fetch_assoc()) {
+            $questionID = $row2["questionID"];
+            DatabaseSessionQuestion::delete($questionID, $mysqli);
+            //Delete from questions
+            $sql = "DELETE FROM `yacrs_questions`
+                        WHERE `questionID` = $questionID";
+            $result = $mysqli->query($sql);
+        }
 
         //Delete from sessionIdentifier
         $sql = "DELETE FROM `yacrs_sessionIdentifier`
                 WHERE `yacrs_sessionIdentifier`.`sessionIdentifier` = $sessionIdentifier";
         $result = $mysqli->query($sql);
 
+        if(!$result) return null;
+
         //Delete from sessions
         $sql = "DELETE FROM `yacrs_sessions`
                 WHERE `sessionID` = $sessionID";
         $result = $mysqli->query($sql);
 
+        if(!$result) return null;
 
-        return $result ? true : false;
+        return true;
     }
 
     /**
